Create new file for Add functionality
