+++ src/dragDrop.js
@@ -4,6 +4,7 @@
 /* eslint-disable no-loop-func */
 
 import { tasksList, items, clearAll } from './index.js';
+import { addTask } from './addTask.js';
 
 export let tasks = [];
 
@@ -52,23 +53,6 @@ export const saveChanges = () => {
       saveLocalstorage(tasks);
     }
   }
-};
-
-const addTask = (task) => {
-  const newTask = document.createElement('div');
-  newTask.classList.add('mini-section');
-  newTask.classList.add('item');
-  newTask.setAttribute('draggable', 'true');
-  newTask.innerHTML = `
-  <span>
-  <input class='check' type='checkbox' id='task-description' name='task-description' value='${task}'>${task}
-  <i class="fas fa-ellipsis-v ellipsis" style="color: gray; float: right;"></i>
-  </span>
-  `;
-  tasksList.appendChild(newTask);
-  saveChanges();
-  saveLocalstorage(tasks);
-  document.location.reload(true);
 };
 
 export const inputListener = (input) => {
